+++ functions/apis/index.js
@@ -2,6 +2,7 @@
  * @file ./functions/apis/index.js
  * @description 公開するCallable APIを集約してexportするエントリーポイントです。
  *****************************************************************************/
+export { checkEmailAvailability } from "./checkEmailAvailability.js";
 export { checkEmailAvailabilityGlobal } from "./checkEmailAvailabilityGlobal.js";
 export { rebuildAllHistories } from "./rebuildAllHistories.js";
 export { rebuildSecurityReportIndexes } from "./rebuildSecurityReportIndexes.js";